+++ services/email/index.js
@@ -1,0 +1,7 @@
+const verifyEmail = require("./verifyEmail");
+const resendVerifyEmail = require("./resendVerifyEmail");
+
+module.exports = {
+  verifyEmail,
+  resendVerifyEmail,
+};
