+++ services/email/resendVerifyEmail.js
@@ -1,0 +1,22 @@
+const {User} = require('../../models')
+
+const resendVerifyEmail = async (req, res) => {
+  const { email } = req.body;
+  const user = await User.findOne({ email });
+  if (!user) {
+    throw new Error("Email not found");
+  }
+  if (user.verify) {
+    throw new Error("Verification has already been passed");
+  }
+  const verifyEmail = {
+    to: email,
+    subject: "Verify email",
+    html: `<a target="_blank" href="${BASE_URL}/api/auth/verify/${user.verificationToken}">Click verify email</a>`,
+  };
+  await sendEmail(verifyEmail);
+
+  res.json({ message: "Verification email sent" });
+};
+
+module.exports = resendVerifyEmail;
